# backend/docs/swagger.yaml
definitions:
  auction.CreateAuctionDTO:
    properties:
      buy_now_price:
        type: integer
      color:
        $ref: '#/definitions/enums.Color'
      date_end:
        type: string
      description:
        type: string
      drive:
        $ref: '#/definitions/enums.Drive'
      engine_capacity:
        type: integer
      engine_power:
        type: integer
      fuel_type:
        $ref: '#/definitions/enums.FuelType'
      manufacturer:
        type: string
      margin:
        $ref: '#/definitions/enums.MarginValue'
      mileage:
        type: integer
      model:
        type: string
      number_of_doors:
        type: integer
      number_of_gears:
        type: integer
      number_of_seats:
        type: integer
      price:
        type: integer
      production_year:
        type: integer
      registration_date:
        type: string
      registration_number:
        type: string
      transmission:
        $ref: '#/definitions/enums.Transmission'
      vin:
        type: string
    required:
    - color
    - description
    - drive
    - engine_capacity
    - engine_power
    - fuel_type
    - manufacturer
    - margin
    - mileage
    - model
    - number_of_doors
    - number_of_gears
    - number_of_seats
    - price
    - production_year
    - registration_date
    - registration_number
    - transmission
    - vin
    type: object
  auction.RetrieveAuctionDTO:
    properties:
      buy_now_price:
        type: integer
      can_modify:
        type: boolean
      color:
        $ref: '#/definitions/enums.Color'
      date_end:
        type: string
      id:
        type: integer
      is_auction:
        type: boolean
      is_liked:
        type: boolean
      main_url:
        type: string
      mileage:
        type: integer
      name:
        type: string
      price:
        type: integer
      production_year:
        type: integer
      status:
        type: string
      username:
        type: string
    type: object
  auction.UpdateAuctionDTO:
    properties:
      buy_now_price:
        type: integer
      color:
        $ref: '#/definitions/enums.Color'
      date_end:
        type: string
      description:
        type: string
      drive:
        $ref: '#/definitions/enums.Drive'
      engine_capacity:
        type: integer
      engine_power:
        type: integer
      fuel_type:
        $ref: '#/definitions/enums.FuelType'
      id:
        type: integer
      manufacturer:
        type: string
      margin:
        $ref: '#/definitions/enums.MarginValue'
      mileage:
        type: integer
      model:
        type: string
      number_of_doors:
        type: integer
      number_of_gears:
        type: integer
      number_of_seats:
        type: integer
      price:
        type: integer
      production_year:
        type: integer
      registration_date:
        type: string
      registration_number:
        type: string
      transmission:
        $ref: '#/definitions/enums.Transmission'
      vin:
        type: string
    type: object
  auth.ChangePasswordDTO:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  auth.ChangePasswordResponse:
    properties:
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
    type: object
  auth.LoginInput:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      company_name:
        type: string
      company_nip:
        type: string
      email:
        type: string
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      person_name:
        type: string
      person_surname:
        type: string
      refresh_token:
        type: string
      selector:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  auth.LogoutInput:
    properties:
      all_devices:
        type: boolean
      refresh_token:
        type: string
    type: object
  auth.RefreshInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RegisterResponse:
    properties:
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
    type: object
...
  custom_errors.HTTPError:
    properties:
      error_description:
        type: string
    type: object
  enums.Color:
    enum:
    - Red
    - Blue
    - Yellow
    - Green
    - Orange
    - Purple
    - Brown
    - Black
    - White
    - Gray
    - Cyan
    - Magenta
    - Lime
    - Navy
    - Teal
    - Maroon
    - Olive
    - Beige
    - Gold
    - Other
    type: string
    x-enum-varnames:
    - RED
    - BLUE
    - YELLOW
    - GREEN
    - ORANGE
    - PURPLE
    - BROWN
    - BLACK
    - WHITE
    - GRAY
    - CYAN
    - MAGENTA
    - LIME
    - NAVY
    - TEAL
    - MAROON
    - OLIVE
    - BEIGE
    - GOLD
    - OTHER
  enums.Drive:
    enum:
    - FWD
    - RWD
    - AWD
    type: string
    x-enum-varnames:
    - FWD
    - RWD
    - AWD
  enums.FuelType:
    enum:
    - Diesel
    - Petrol
    - Electric
    - Ethanol
    - LPG
    - Biofuel
    - Hybrid
    - Hydrogen
    type: string
    x-enum-varnames:
    - DIESEL
    - PETROL
    - ELECTRIC
    - ETHANOL
    - LPG
    - BIOFUEL
    - HYBRID
    - HYDROGEN
  enums.MarginValue:
    enum:
    - 3
    - 5
    - 10
    type: integer
    x-enum-varnames:
    - LOW_MARGIN
    - MEDIUM_MARGIN
    - HIGH_MARGIN
  enums.Transmission:
    enum:
    - Manual
    - Automatic
    - Cvt
    - Dual clutch
    type: string
    x-enum-varnames:
    - MANUAL
    - AUTOMATIC
    - CVT
    - DUAL_CLUTCH
...
  user.CreateUserDTO:
    properties:
      company_name:
        type: string
      company_nip:
        type: string
      email:
        type: string
      password:
        type: string
      person_name:
        type: string
      person_surname:
        type: string
      selector:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - selector
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    name: BD2
  description: Car-Dealer API
  license:
    name: MIT
  title: Car‑Dealer API
  version: "1.0"
paths:
  /auction:
    get:
      consumes:
      - application/json
      description: Retrieves all available auctions from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/auction.RetrieveAuctionDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get all auctions
      tags:
      - auction
    post:
      consumes:
      - application/json
      description: Creates a new auction with the provided details
      parameters:
      - description: Auction details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auction.CreateAuctionDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created auction
          schema:
            $ref: '#/definitions/auction.RetrieveAuctionDTO'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Create Auction
      tags:
      - auction
    put:
      consumes:
      - application/json
      description: Updates an existing auction with the provided details
      parameters:
      - description: Auction details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auction.UpdateAuctionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated auction
          schema:
            $ref: '#/definitions/auction.RetrieveAuctionDTO'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Update auction
      tags:
      - auction
  /auction/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific auction by its ID
      parameters:
      - description: Auction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Delete auction by ID
      tags:
      - auction
    get:
      consumes:
      - application/json
      description: Retrieves a specific auction by its ID
      parameters:
      - description: Auction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auction.RetrieveAuctionDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get auction by ID
      tags:
      - auction
  /auction/buy-now/{id}:
    post:
      consumes:
      - application/json
      description: Allows a user to instantly purchase an auction at its buy now price
        if available
      parameters:
      - description: Auction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully purchased the auction
        "400":
          description: Invalid auction ID or buy now operation failed
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized - user not logged in
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Buy an auction at its buy now price
      tags:
      - auction
  /auth/change-password:
    put:
      consumes:
      - application/json
      description: Changes the password of the authenticated user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Password change details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ChangePasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Password successfully changed
        "400":
          description: Invalid request or password validation failed
          schema:
            $ref: '#/definitions/auth.ChangePasswordResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/auth.ChangePasswordResponse'
      security:
      - Bearer: []
      summary: Change user password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a pair of tokens and user data
      parameters:
      - description: Login form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK - returns tokens and data of the user
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.LoginResponse'
      security:
      - Bearer: []
      summary: Login user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access tokens
      parameters:
      - description: Refresh token form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK - returns new access token
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Set up account and return the status
      parameters:
      - description: Registration form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created - returns tokens
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "409":
          description: Login taken
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
      summary: Register new user
      tags:
      - auth
...
  /logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate refresh token
      parameters:
      - description: Logout form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LogoutInput'
      produces:
      - application/json
      responses:
        "204":
          description: No content
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Logout user
      tags:
      - auth
...
schemes:
- http
swagger: "2.0"